+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.4.0
+
+- Add `processUrl()` to handle website URLs
+- Add canonical links
+- Add RSS alternate link
+- Update NPM packages
+
 ## 1.3.5
 
 - Add documentation for setting up [live previews](./README.md#live-previews)